Created RockPaperScissorsFrame and computerLogic
Created GUI with 4 panels, and outer panel and a top/mid/bottom panel contaiend within the outer panel. Game results are JLabels in the topPanel,  the text area/scroller is in the midPanel, and the buttons to play with are in the botPanel. The function computerLogic determines a random move for the computer and updates the result panels accordingly while printing out the events of the game as it occurs.